+++ src/chunk_metadata.rs
@@ -27,17 +27,17 @@ fn parse_prefixes<T: NumberLike>(
   reader: &mut BitReader,
   flags: &Flags,
 ) -> QCompressResult<Vec<Prefix<T>>> {
-  let n_pref = reader.read_usize(MAX_COMPRESSION_LEVEL as usize)?;
+  let n_pref = reader.read_usize(MAX_COMPRESSION_LEVEL)?;
   let mut prefixes = Vec::with_capacity(n_pref);
   let bits_to_encode_prefix_len = flags.bits_to_encode_prefix_len();
   for _ in 0..n_pref {
-    let count = reader.read_usize(BITS_TO_ENCODE_N_ENTRIES as usize)?;
+    let count = reader.read_usize(BITS_TO_ENCODE_N_ENTRIES)?;
     let lower = T::read_from(reader)?;
     let upper = T::read_from(reader)?;
     let code_len = reader.read_usize(bits_to_encode_prefix_len)?;
     let val = reader.read(code_len)?;
     let jumpstart = if reader.read_one()? {
-      Some(reader.read_usize(BITS_TO_ENCODE_JUMPSTART as usize)?)
+      Some(reader.read_usize(BITS_TO_ENCODE_JUMPSTART)?)
     } else {
       None
     };
@@ -47,10 +47,10 @@ fn parse_prefixes<T: NumberLike>(
 }
 
 fn write_prefixes<T: NumberLike>(prefixes: &[Prefix<T>], writer: &mut BitWriter, flags: &Flags) {
-  writer.write_usize(prefixes.len(), MAX_COMPRESSION_LEVEL as usize);
+  writer.write_usize(prefixes.len(), MAX_COMPRESSION_LEVEL);
   let bits_to_encode_prefix_len = flags.bits_to_encode_prefix_len();
   for pref in prefixes {
-    writer.write_usize(pref.count, BITS_TO_ENCODE_N_ENTRIES as usize);
+    writer.write_usize(pref.count, BITS_TO_ENCODE_N_ENTRIES);
     pref.lower.write_to(writer);
     pref.upper.write_to(writer);
     writer.write_usize(pref.val.len(), bits_to_encode_prefix_len);
@@ -61,7 +61,7 @@ fn write_prefixes<T: NumberLike>(prefixes: &[Prefix<T>], writer: &mut BitWriter,
       },
       Some(jumpstart) => {
         writer.write_one(true);
-        writer.write_usize(jumpstart, BITS_TO_ENCODE_JUMPSTART as usize);
+        writer.write_usize(jumpstart, BITS_TO_ENCODE_JUMPSTART);
       },
     }
   }
@@ -69,8 +69,8 @@ fn write_prefixes<T: NumberLike>(prefixes: &[Prefix<T>], writer: &mut BitWriter,
 
 impl<T> ChunkMetadata<T> where T: NumberLike {
   pub fn parse_from(reader: &mut BitReader, flags: &Flags) -> QCompressResult<Self> {
-    let n = reader.read_usize(BITS_TO_ENCODE_N_ENTRIES as usize)?;
-    let compressed_body_size = reader.read_usize(BITS_TO_ENCODE_COMPRESSED_BODY_SIZE as usize)?;
+    let n = reader.read_usize(BITS_TO_ENCODE_N_ENTRIES)?;
+    let compressed_body_size = reader.read_usize(BITS_TO_ENCODE_COMPRESSED_BODY_SIZE)?;
     let prefix_info = if flags.delta_encoding_order == 0 {
       let prefixes = parse_prefixes::<T>(reader, flags)?;
       PrefixInfo::Simple {
